requirements.txt
Flask~=2.0.1
google-auth~=1.28.0
requests~=2.23.0
